[player][inputstream][addons] force to use VideoPlayer by inputstream addon
Without it check only the file item for Audio, Video or Game and mostly
(until now all) has used only videostream via a inputstream addon.
As a result, this error was never noticed.

But in the event that only one audio stream is available, this falls back
to the PAPlayer which cannot be used with the input stream addon.

This checks whether the requested file / stream is handled by an inputstream
addon and if so only the VideoPlayer is used.